+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-0.0.2 (unreleased)
-------------------
+0.1.0
+-----
 - Added replacing parameters in translation strings
 - Added key prefix
 - Removed autoloader and debug stuff